requirements.txt
asgiref==3.7.2
boto3==1.26.142
botocore==1.29.142
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.7
defusedxml==0.8.0rc2
dj-database-url==2.0.0
dj-rest-auth==5.0.2
Django==4.2.7
django-allauth==0.59.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.2
django-heroku==0.3.1
django-storages==1.13.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djongo==1.3.6
djoser==2.2.2
dnspython==2.4.2
drf-yasg==1.21.7
environs==14.3.0
gunicorn==23.0.0
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
marshmallow==4.0.0
oauthlib==3.2.2
packaging==23.2
paystackapi==2.1.1
phonenumbers==8.13.26
Pillow==10.1.0
psycopg2==2.9.6
psycopg2-binary==2.9.10
pycparser==2.21
PyJWT==2.8.0
pymongo==4.5.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.1.1
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.6.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
typing_extensions==4.6.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
whitenoise==6.4.0
